#1 Fix incorrect project title during confirmation
The previously used #project_title method at this
point always returns "lingohub_ruby". This is
obviously not correct, and will have to be fixed
as well. This commit just makes sure that we don't
confuse the other with having to enter a wrong
project name. Also it seems to me that it's totally
fine to actually just use project.title in there
anyway.